Split Wayback archive library modules
