+++ loom/wayback_archive/BUILD.bazel
@@ -11,7 +11,19 @@ exports_files([
 rust_library(
     name = "wayback_archive_lib",
     srcs = [
+        "ia_client.rs",
         "lib.rs",
+        "limiter.rs",
+        "metrics.rs",
+        "path.rs",
+        "response.rs",
+        "service.rs",
+        "store/memory.rs",
+        "store/mod.rs",
+        "store/postgres.rs",
+        "store/s3.rs",
+        "types.rs",
+        "util.rs",
     ],
     crate_name = "wayback_archive",
     edition = "2024",
