+++ loom/wayback_archive/util.rs
@@ -1,0 +1,14 @@
+use std::fmt::Write;
+
+use sha2::{Digest, Sha256};
+
+pub(crate) fn sha256_hex(body: &[u8]) -> String {
+    let mut hasher = Sha256::new();
+    hasher.update(body);
+    let digest = hasher.finalize();
+    let mut encoded = String::with_capacity(64);
+    for byte in digest {
+        write!(&mut encoded, "{byte:02x}").expect("writing to String should not fail");
+    }
+    encoded
+}
